Move OpenSSL local tests to validation
