--- certs/ca.rsa4096.crt
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFyTCCA7GgAwIBAgIUYwbIsHSBxdFxuzRJnBbyNIkdDZMwDQYJKoZIhvcNAQEL
-BQAwdDELMAkGA1UEBhMCWFgxCzAJBgNVBAgMAllZMRMwEQYDVQQHDApBbnRhcmN0
-aWNhMRMwEQYDVQQKDApSdXN0Q3J5cHRvMRUwEwYDVQQLDAxDb250cmlidXRvcnMx
-FzAVBgNVBAMMDmNhLnJ1c3RjcnlwLnRvMB4XDTI0MDQyNDE1NTAzOVoXDTI0MDUy
-NDE1NTAzOVowdDELMAkGA1UEBhMCWFgxCzAJBgNVBAgMAllZMRMwEQYDVQQHDApB
-bnRhcmN0aWNhMRMwEQYDVQQKDApSdXN0Q3J5cHRvMRUwEwYDVQQLDAxDb250cmli
-dXRvcnMxFzAVBgNVBAMMDmNhLnJ1c3RjcnlwLnRvMIICIjANBgkqhkiG9w0BAQEF
-AAOCAg8AMIICCgKCAgEAvU2Mig6qXtYz9Ga5xhH3+jQE/QBcUPoSlFBu699yT+Zk
-SARUung1ZR0p8w4EtP8g4Avb3YoKazd3LlcCBQDtRT/9NTa8Wz6cp/d7OgZWvGeE
-W10EhKIOm8beBSWYi0qdi6xnXBMJEirx958lH/v+zXLoKUXmRWbYjdd84igLVvTW
-bh4UDg36qrVN7zYK6XjI/k3khvAqZh5/wkK9XIAlwWTItgypJjCXxlcNsM55D0UV
-PLKUflaZIfyc2e44k9T3ZTFE/4r5KVmCiPzPhJDFhkXvbVQFLnGYlJ+Sz+HbAEM+
-WE10DdGbPu2SWnTkh3qIYJmewWGuynca//h/bX/tdyeWOwP6GNZgMBSw/Opz/95a
-xFK/WZ5f6FyIU5K3XvuaFnQy4LaecQBn43EV+Dh6ATByDSAm058cHaon3FOIrR8T
-l07ugDe6AEjjO4SzM9NwaxykN6p0ig8Hp+kReVx8Lr/MBf0L6oY00LD9pzP3h2+A
-Tuzwi4+2Z/xhDOm8f/tzyLSB7wFrlCm3CP3MNel/aV+P1vSKsdMNHaHNJpwgVsjI
-VBjchXKkJHH8NxEQoox/9lB3B7YhwfZ9Kxmce+8qaC6vejEJ00bYqfsNcsTDNAHU
-mUyJfswBD6hWkPLVazjbRVqCzHPk46Lmyzgx+uKyisrF6rrch98hcAiWOMh/OkUC
-AwEAAaNTMFEwHQYDVR0OBBYEFAIr9ZdW0YZXqmSE1HBF0SFbATwOMB8GA1UdIwQY
-MBaAFAIr9ZdW0YZXqmSE1HBF0SFbATwOMA8GA1UdEwEB/wQFMAMBAf8wDQYJKoZI
-hvcNAQELBQADggIBAJA+cnD6gTsg/PHmrXXc8/Psa8RffrwFmnldV23fJECe011K
-GPNI/GMszCPMeV1EVHfhxNulYBMfdacfNSr14TI2y1ErCwFfISRCNPLEDHX5JqVE
-HHuCIKYq6XV3tHX4bLcq7qprQp3bq7LaqYcktfpN4eoxSwNRYPZpMXV1OQepl5jM
-wIvP7kWkeCQTFZ/5xt0rWvqnYZ4bQMXB7wqsty7oYzoivsUcpgzeUnmDO9Ad0g1q
-QsW8hcLxwChzcV+DtHjX1d1hP6NbZtlUOkmTwYMWdx6+8IFG6WeIMfjouXqNI4fj
-5R/IJ+KqrRMBsRagDR4PPnIZEb21g/0nqF+YKGrIyiUu2+4yN8R/qXFQfQ5yYoB5
-d4P6PokqOQhyNq8oAKYK0iMU8Ju9mG2CjjS/W6HKqVfaE80foZBs0ohGZ7eh045Y
-pCQh/Bwi5StECKk82TXHSQNjZaWlQsXYbVHM/uoJ6Rh38Lmn9CqjhRH9dKkOcMP9
-NWmDAnTnTko37li0Sr92hZoyKc77MNYBp15KlKGA3B+dPqu0pgMznVK0B1ddt9z3
-/hRHR6YSpULAB9NelAk8r556Sx6OFM9hC4i3OewJwoJBtyIrBHd1UEkq43FUuzug
-DeOzb/dT46OKlXNdLK1Idpeh92f29vV0duHqJCd2HQ1OvMHlPY0c77Cpq8SC
------END CERTIFICATE-----
