--- certs/rustcryp.to.rsa4096.ca_signed.crt
@@ -1,38 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIGszCCBJugAwIBAgIUbq8paFssW4RHGb6JM15HbfQgTFYwDQYJKoZIhvcNAQEL
-BQAwdDELMAkGA1UEBhMCWFgxCzAJBgNVBAgMAllZMRMwEQYDVQQHDApBbnRhcmN0
-aWNhMRMwEQYDVQQKDApSdXN0Q3J5cHRvMRUwEwYDVQQLDAxDb250cmlidXRvcnMx
-FzAVBgNVBAMMDmNhLnJ1c3RjcnlwLnRvMB4XDTI0MDQyNDE1NTAzOVoXDTI1MDQy
-NDE1NTAzOVowgZUxCzAJBgNVBAYTAlhYMQswCQYDVQQIDAJZWTETMBEGA1UEBwwK
-QW50YXJjdGljYTEVMBMGA1UECgwMQ29udHJpYnV0b3JzMRAwDgYDVQQLDAdUZXN0
-ZXJzMSAwHgYJKoZIhvcNAQkBFhFoZWxsb0BydXN0Y3J5cC50bzEZMBcGA1UEAwwQ
-dGVzdC5ydXN0Y3J5cC50bzCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIB
-AMV97RVGkaTvpNZKhzBAFVU1iRBtnESqnQapIQmuv/vxn/b0ldgCc3E44UATupzp
-THurOIY7f/lyc4nzafsLYhFpYySsRpGodL9sHddM8RBKaTM63eqkLU+pegb4f0Hf
-YISfdWdEYScgRMeDfp/UhAOU7QMsJp0bywic0qLI+tJSjEEPz974lbEfD2wjDLtg
-+VmPzW4AD1TlbTk7tJqkG7T6vqC3772rPfVE7bu4lvX4YTFDSHbX0/rue/QnebzC
-qrLkSZHFhb48sc6vPsdAOKTzKeyAEcYpy4yqRy746k1yFGW/2o2VDdEQKtqX0mlH
-V/UqPfjWq2JmnKQr0+0sTLp0MPU39hazeng0vQ7q9gCK7pV+InRpUtR13EjH0J1d
-WDLXrNmrdnVOudMAbaC5C6JGMZMmxuQs9GiuAKf/SRj9R59GIKdhN42O1m0oGB/e
-xj2VFXxevVldEpuauA21qfcYnO/8NrCDzp+QLX56cnhScH+o6V1K4B8njjD8Ehdc
-cLZzMaW62RzaFwx5+GCAF9HeoMiBXk/2QGVyqhFLPTSAJ7BEqPKEusCS9cuIgMrU
-PPfBKEVboRj4S1e75Z98GeuBKKP1UfMvTEGbOww8FdYE1FPSqjTIhlZZ/Lnstzc6
-vBqtSy8ghhh7eBsJ39MQ8FBrCVwysTQjNqv+RiHZO6SDAgMBAAGjggEZMIIBFTAM
-BgNVHRMBAf8EAjAAMAsGA1UdDwQEAwIGwDAdBgNVHQ4EFgQUQy01BMF7zjnlrQAc
-4/D00DU0vr0wgbEGA1UdIwSBqTCBpoAUAiv1l1bRhleqZITUcEXRIVsBPA6heKR2
-MHQxCzAJBgNVBAYTAlhYMQswCQYDVQQIDAJZWTETMBEGA1UEBwwKQW50YXJjdGlj
-YTETMBEGA1UECgwKUnVzdENyeXB0bzEVMBMGA1UECwwMQ29udHJpYnV0b3JzMRcw
-FQYDVQQDDA5jYS5ydXN0Y3J5cC50b4IUYwbIsHSBxdFxuzRJnBbyNIkdDZMwJQYD
-VR0RBB4wHIIPd3d3LnJ1c3RjcnlwLnRvgglsb2NhbGhvc3QwDQYJKoZIhvcNAQEL
-BQADggIBAISod7roIUqts+9TWQAwBXt3pNaEGKLVabIN+AxeaP/1kPoefZV/VqUA
-kqEpQLH9Yfhe7rKtGAKYYoZmjePmNHoo8aIUXLG9HVqCDwBKZ/at69GdEZMIoP0r
-lq+v0inMJ4q/mrgc9GLq7gyvLP97qPAQRlhIoCioT7lzBNxn4+mZ/8rlMfKq4gqQ
-7B2GVtgA1G/EIt4JCZdnjieWz0/+HoQBD9MuTwYilBYarkAgdfvFKFNLlL9oSjVp
-6CYJZGhzODQKxUy6MABgsvGtOkZPwnWPO6VBeR8RBFF09npjTk7W0Pr47Q0Vkd82
-lSvoGD1B0vme7baJVvzJKCyyJVGraZ/1aJftS2t+8A3XIeajgnoLp+Glhi8gVT3o
-URAA4WTIFnqkswyPS2rzttChuNUwRjm2GhrTwD3uEri16NV8mjilL3CbxJhjbUaB
-pX1D4phoEvGMudK9aXI9UXYaa0BRbI6ha6lQoKlMk7LG9zaroEtakaPcJwYCKufG
-CZ2FK0e+vwnNDwOD5gWHCKHxCrWaLX5obZ4oKvl5rEw+R7/YiSySBEqrCK9aLgW1
-M7Yg8C8wUhW+3MtoQBrvOcODpNDfnOgkXaWlKZVY3PB2SN+2ICFsnrsSOK98lC6v
-bIZ7Wf0o6/As83B9HBqvlg2wx20DVpnHZU7cZ2Iy/VSYxP8qirxd
------END CERTIFICATE-----
